third_party/WebKit/LayoutTests/plugins/get-user-agent-with-null-npp-from-npp-new-expected.txt

CONSOLE MESSAGE: line 7: PLUGIN: SUCCESS!

This tests that we can get a user agent when calling NPN_UserAgent with a null NPP from NPP_New.
